Decided to move them back into an ffi dir but kept the separate library
