+++ pysrc/otto/__init__.py
@@ -1,0 +1,1 @@
+import otto_ffi.ld_path
