--- pysrc/otto_ffi/__main__.py
@@ -1,3 +1,0 @@
-from otto.ffi import build
-
-build()
