+++ setup.cfg
@@ -21,8 +21,9 @@ classifiers =
 [options]
 packages =
     otto_ffi
-    otto_ffi.cl
     otto_ffi.config
+    otto_ffi.ffi
+    otto_ffi.ffi.cl
     otto_utils
 python_requires = >=3.7
 package_dir =
